git-svn-id: https://svn.neumont.edu:8443/svn/taylorgarciarepspot@175 9283bf75-a890-104e-bf60-28af5b2e9440

+++ GameSkeleton/Engine2.0/Time/Clock.cpp
@@ -11,7 +11,7 @@ void Clock::stop()
 	QueryPerformanceCounter(&thisTime);
 	LARGE_INTEGER delta;
 	delta.QuadPart = thisTime.QuadPart - lastStartTime.QuadPart;
-	deltaTime = ((float)delta.QuadPart) / timeFrequency.QuadPart;
+	deltaTime = ((float)delta.QuadPart) / ((float)timeFrequency.QuadPart);
 	deltaLastLap.QuadPart = thisTime.QuadPart;
 }
 void Clock::lap()
